--- Realworld/src/goat_control/goat_control/core/control/__init__.py
@@ -1,26 +1,0 @@
-# goat_control/core/control/__init__.py
-from .pd_controller import PDJointController, PDControllerConfig
-from .pi_controller import WheelPIController, WheelPIControllerConfig
-from .safety_limiter import (
-    TorqueSafetyLimiter,
-    SafetyLimiterConfig,
-    ConditionalIntegratorAntiWindup,
-    ConditionalIntegratorConfig,
-)
-
-__all__ = [
-    # IK
-    "InverseKinematicsSolver",
-    "IKResult",
-    "IkMode",
-    # Controllers
-    "PDJointController",
-    "PDControllerConfig",
-    "WheelPIController",
-    "WheelPIControllerConfig",
-    # Safety / Anti-windup
-    "TorqueSafetyLimiter",
-    "SafetyLimiterConfig",
-    "ConditionalIntegratorAntiWindup",
-    "ConditionalIntegratorConfig",
-]
