--- Realworld/src/goat_control/goat_control/core/model/__init__.py
@@ -1,14 +1,0 @@
-# goat_control/core/model/__init__.py
-from .goat_model import GoatModel, GoatModelConfig, EffortOutputMode
-from .model_builder import (
-    build_goat_model_from_yaml,
-    build_control_pipeline_from_yaml,
-)
-
-__all__ = [
-    "GoatModel",
-    "GoatModelConfig",
-    "EffortOutputMode",
-    "build_goat_model_from_yaml",
-    "build_control_pipeline_from_yaml",
-]
