+++ Realworld/src/goat_control/goat_control/nodes/calibration_node.py
@@ -2,7 +2,8 @@
 import rclpy
 from rclpy.node import Node
 from sensor_msgs.msg import JointState
-from motor_interfaces.msg import BaseStates
+from goat_control.utils.imu.quaternion_utils import *
+from motor_interfaces.msg import ImuState
 
 import yaml
 import os
@@ -26,11 +27,18 @@ class CalibrationNode(Node):
 
         # Subscriber
         self.joint_state_subscriber = self.create_subscription(
-            JointState, "joint_states", self._on_joint_state_msg, 10
+            JointState, "/joint_states", self._on_joint_state_msg, 10
         )
         self.imu_subscriber = self.create_subscription(
-            BaseStates, "/goat/imu_data", self._on_imu_msg, 10
+            ImuState, "/imu", self._on_imu_msg, 10
         )
+
+        # YAML file
+        with open(self.yaml_path, 'r', encoding='utf-8') as file_handle:
+            self.cfg = yaml.safe_load(file_handle)
+
+        self.old_joint_offsets = self.cfg["joint_offsets"]
+        self.old_imu_offsets = self.cfg["imu_offsets"]
 
         # Data buffers
         self.latest_joint_state = None
@@ -57,7 +65,7 @@ class CalibrationNode(Node):
     def _on_joint_state_msg(self, msg: JointState):
         self.latest_joint_state = msg
 
-    def _on_imu_msg(self, msg: BaseStates):
+    def _on_imu_msg(self, msg: ImuState):
         self.latest_imu_state = msg
 
     def _get_key(self):
@@ -134,32 +142,86 @@ class CalibrationNode(Node):
             if joint_names is None:
                 joint_names = self.latest_joint_state.name
             
-            # 1. Store current positions
+            # Store current positions
             current_pos = np.array(self.latest_joint_state.position, dtype=float)
+            current_pos += self.old_joint_offsets                                         # Restore original position
             position_samples.append(current_pos)
             
             # Wait for next update
             time.sleep(sleep_interval)
 
-        # 2. Calculate Average
+        # Calculate Average
         avg_positions = np.mean(position_samples, axis=0)
 
-        # 3. Calculate offsets (Average)
+        # Calculate offsets (Average)
         joint_offsets = avg_positions
 
         self.get_logger().info(f"Averaged Positions ({self.sample_count} samples): {avg_positions}")
         self.get_logger().info(f"Calculated Joint Offsets: {joint_offsets}")
 
-        # 4. Save to YAML
+        # Save to YAML
         self._save_joint_offsets_to_yaml(joint_offsets)
     
     def _imu_calibration(self):
         """IMU Calibration Logic (Placeholder)."""
-        # TODO: Implement IMU calibration logic here
-        pass
+        # Settings for sampling
+        sleep_interval = 0.05  # 20 * 0.05 = 1.0 second total duration
+
+        self.get_logger().info(f"Collecting {self.sample_count} samples (approx 1 sec)... Keep robot still.")
+        
+        # Joint position buffer list
+        quat_samples = []
+
+        # Sampling Loop
+        for i in range(self.sample_count):
+            
+            # Exception
+            if self.latest_imu_state is None:
+                self.get_logger().warn("IMU data lost during sampling!")
+                return
+            
+            # Store current positions
+            current_quat = np.array(self.latest_imu_state.quat, dtype=float)
+            old_joint_offsets_inv = inverse_quat(self.old_joint_offsets)
+            current_quat = multiply_quat(old_joint_offsets_inv, current_quat)               # Restore original quaternion
+            quat_samples.append(current_quat)
+            
+            # Wait for next update
+            time.sleep(sleep_interval)
+
+        # Calculate Average
+        avg_quat = np.mean(quat_samples, axis=0)
+        avg_quat /= np.linalg.norm(avg_quat)
+
+        z_axis_local = np.array([0.0, 0.0, 1.0])
+        v_up = rotate_vector_by_quat(avg_quat, z_axis_local)
+
+        # Target Z axis
+        v_target = np.array([0.0, 0.0, 1.0])
+
+        
+        axis = np.cross(v_up, v_target)
+        axis_norm = np.linalg.norm(axis)
+
+        # Already robot is upright
+        if axis_norm < 1e-8:
+            quat_offsets = np.array([1.0, 0.0, 0.0, 0.0])
+
+        axis = axis / axis_norm
+        dot_prod = np.clip(np.dot(v_up, v_target), -1.0, 1.0)
+        angle = np.arccos(dot_prod)
+
+        # Offset quaternion
+        quat_offsets = axis_angle_to_quat(axis, angle)
+
+        self.get_logger().info(f"Calculated IMU quaternion Offsets: {quat_offsets}")
+
+        # Save to YAML
+        self._save_imu_offsets_to_yaml(quat_offsets)
+        
 
     def _save_joint_offsets_to_yaml(self, offsets):
-        # 1. Read existing file
+        # Read existing file
         data = {}
         if os.path.exists(self.yaml_path):
             try:
@@ -169,15 +231,15 @@ class CalibrationNode(Node):
                 self.get_logger().error(f"Failed to load existing YAML: {e}")
                 return
         
-        # 2. Update data structure
+        # Update data structure
         formatted_offsets = [float(val) for val in offsets]
         
-        if 'calibration' not in data:
-            data['calibration'] = {}
+        if 'joint_offsets' not in data:
+            data['joint_offsets'] = {}
             
-        data['calibration']['joint_offsets'] = formatted_offsets
+        data['joint_offsets'] = formatted_offsets
 
-        # 3. Write to file
+        # Write to file
         try:
             os.makedirs(os.path.dirname(self.yaml_path), exist_ok=True)
             with open(self.yaml_path, 'w') as f:
@@ -189,7 +251,7 @@ class CalibrationNode(Node):
             self.get_logger().error(f"Failed to write YAML file: {e}")
 
     def _save_imu_offsets_to_yaml(self, offsets):
-        # 1. Read existing file
+        # Read existing file
         data = {}
         if os.path.exists(self.yaml_path):
             try:
@@ -199,15 +261,15 @@ class CalibrationNode(Node):
                 self.get_logger().error(f"Failed to load existing YAML: {e}")
                 return
         
-        # 2. Update data structure
+        # Update data structure
         formatted_offsets = [float(val) for val in offsets]
         
-        if 'calibration' not in data:
-            data['calibration'] = {}
+        if 'imu_offsets' not in data:
+            data['imu_offsets'] = {}
             
-        data['calibration']['imu_offsets'] = formatted_offsets
+        data['imu_offsets'] = formatted_offsets
 
-        # 3. Write to file
+        # Write to file
         try:
             os.makedirs(os.path.dirname(self.yaml_path), exist_ok=True)
             with open(self.yaml_path, 'w') as f:
