+++ Realworld/src/goat_control/goat_control/nodes/log_viewer_node.py
@@ -6,17 +6,17 @@ from typing import List, Optional
 
 import numpy as np
 import rclpy
+import yaml
 from rclpy.node import Node
-from std_msgs.msg import Float32MultiArray
+from message_filters import Subscriber, ApproximateTimeSynchronizer
 from sensor_msgs.msg import JointState
-
 
 @dataclass
 class LatestLog:
     vector: Optional[np.ndarray] = None
 
 
-class LogViewer(Node):
+class LogViewerNode(Node):
     """
     Subscribe:  goat/torque_log (Float32MultiArray)
       - data (supported layouts):
@@ -32,159 +32,128 @@ class LogViewer(Node):
     """
 
     def __init__(self):
-        super().__init__("log_viewer")
+        super().__init__("log_viewer_node")
 
-        # Params
-        self.declare_parameter("log_topic", "goat/torque_log")
-        self.declare_parameter("joint_state_topic", "joint_states")
-        self.declare_parameter("use_joint_state_names", True)
+        # Parameters
+        self.declare_parameter("yaml_path", "src/goat_control/config/goat_config.yaml")
+        self.declare_parameter("sample_count", 20)
 
-        self.declare_parameter("num_joints", 8)
-        self.declare_parameter(
-            "joint_names",
-            ["hip_L",
-             "hip_R",
-             "thigh_L",
-             "thigh_R",
-             "knee_L", 
-             "knee_R",
-             "wheel_L",
-             "wheel_R"]
-        )
+        # YAML file
+        yaml_path = str(self.get_parameter("yaml_path").value)
 
-        # For nicer ref printing (ref is position for joints, speed for wheels)
-        self.declare_parameter("wheel_indices", [6, 7])
+        with open(yaml_path, "r", encoding="utf-8") as file_handle:
+            self.cfg = yaml.safe_load(file_handle)
 
         self.declare_parameter("print_rate_hz", 50.0)
         self.declare_parameter("print_degrees", True)
-        self.declare_parameter("command_unit", "torque_nm")  # torque_nm or amp
         self.declare_parameter("precision", 3)
-
-        self.log_topic = str(self.get_parameter("log_topic").value)
-        self.joint_state_topic = str(self.get_parameter("joint_state_topic").value)
-        self.use_joint_state_names = bool(self.get_parameter("use_joint_state_names").value)
-
-        self.num_joints = int(self.get_parameter("num_joints").value)
-        self.joint_names: List[str] = [str(x) for x in self.get_parameter("joint_names").value]
-        self.wheel_indices = [int(x) for x in self.get_parameter("wheel_indices").value]
 
         self.print_rate_hz = float(self.get_parameter("print_rate_hz").value)
         self.print_degrees = bool(self.get_parameter("print_degrees").value)
-        self.command_unit = str(self.get_parameter("command_unit").value)
         self.precision = int(self.get_parameter("precision").value)
+        
+        # YAML parameters
+        self.num_joints = self.cfg["num_joints"]
+        self.joint_names = self.cfg["joint_names"]
+        self.wheel_indices = self.cfg["wheel_indices"]
+        self.gear_ratio = np.array(self.cfg["motor_gear_ratio"], dtype=float)
 
-        if len(self.joint_names) != self.num_joints:
-            self.get_logger().warn(
-                f"joint_names length ({len(self.joint_names)}) != num_joints ({self.num_joints}). "
-                "Falling back to generic names."
-            )
-            self.joint_names = [f"joint_{i}" for i in range(self.num_joints)]
-
-        self.latest = LatestLog()
+        self.joint_current: Optional[JointState] = None
+        self.joint_ref: Optional[JointState] = None
         self._print_count = 0
+        self.command_unit = "Nm"
 
-        # Subs
-        self.create_subscription(Float32MultiArray, self.log_topic, self._on_log, 10)
-
-        if self.use_joint_state_names:
-            self.create_subscription(JointState, self.joint_state_topic, self._on_joint_state, 10)
+        # Subscribers
+        self.create_subscription(JointState, "/commands", self._on_joint_ref, 10)
+        self.create_subscription(JointState, "/joint_states", self._on_joint_state, 10)
+        # self.command_subscriber = Subscriber(self, JointState, "/commands", 10)
+        # self.state_subscriber = Subscriber(self, JointState, "/joint_states", 10)
+        # self.time_sync = ApproximateTimeSynchronizer([self.command_subscriber, self.state_subscriber], 10, 0.01)
+        # self.time_sync.registerCallback(self._sync_callback)
 
         # Timer (rate-limit printing)
         period_sec = 1.0 / max(self.print_rate_hz, 0.5)
         self.create_timer(period_sec, self._tick)
 
         self.get_logger().info(
-            f"MotorTorqueLogViewer started. Subscribing '{self.log_topic}' "
-            f"(names from JointState: {self.use_joint_state_names})."
+            "LogViewerNOde started."
+            f"(names of Joints: {self.joint_names})."
         )
 
-    def _on_log(self, msg: Float32MultiArray) -> None:
-        data = np.asarray(msg.data, dtype=float).flatten()
-        self.latest.vector = data
+    def _on_joint_ref(self, msg: JointState) -> None:
+        self.joint_ref = msg
 
     def _on_joint_state(self, msg: JointState) -> None:
-        # Use incoming names if they look valid and match num_joints
-        if msg.name and len(msg.name) == self.num_joints:
-            self.joint_names = list(msg.name)
+        self.joint_current = msg
+
+    def _sync_callback(self, msg_ref, msg_current):
+        self._on_joint_ref(msg_ref)
+        self._on_joint_state(msg_current)
 
     def _tick(self) -> None:
-        if self.latest.vector is None:
+        # No subscription
+        if self.joint_current is None:
             return
+        
+        if self.joint_ref is None:
+            self.joint_ref = self.joint_current
 
-        vector = self.latest.vector
-
-        # Layout parsing (ONLY accept 4N, 3N support removed)
-        expected_len = 5 * self.num_joints
-        if vector.size != expected_len:
-            self.get_logger().warn(
-                f"log length mismatch: got {vector.size}, expected {expected_len}"
-            )
-            return
-
-        # Data slicing
-        joint_position_rad = vector[0 : self.num_joints]
-        joint_velocity_rad_per_sec = vector[self.num_joints : 2 * self.num_joints]
-        command_value = vector[2 * self.num_joints : 3 * self.num_joints]
-        safe_joint_targets = vector[3 * self.num_joints : 4 * self.num_joints]
-        ref_vector = vector[4 * self.num_joints : 5 * self.num_joints]
-
+        # Decode JointState msg
         if self.print_degrees:
-            joint_position = np.rad2deg(joint_position_rad)
-            joint_velocity = np.rad2deg(joint_velocity_rad_per_sec)
+            joint_pos_current = np.rad2deg(self.joint_current.position)
+            joint_vel_current = np.rad2deg(self.joint_current.velocity)
+            joint_pos_ref = np.rad2deg(self.joint_ref.position)
+            joint_vel_ref = np.rad2deg(self.joint_ref.velocity)
             position_unit = "deg"
             velocity_unit = "deg/s"
         else:
-            joint_position = joint_position_rad
-            joint_velocity = joint_velocity_rad_per_sec
+            joint_pos_current = np.array(self.joint_current.position, dtype=float)
+            joint_vel_current = np.array(self.joint_current.velocity, dtype=float)
+            joint_pos_ref = np.array(self.joint_ref.position, dtype=float)
+            joint_vel_ref = np.array(self.joint_ref.velocity, dtype=float)
             position_unit = "rad"
             velocity_unit = "rad/s"
 
-        command_unit = "Nm" if self.command_unit == "torque_nm" else "A"
+        joint_effort_ref = np.array(self.joint_ref.effort, dtype=float)
+        joint_effort_current = np.array(self.joint_current.effort, dtype=float)
+
+        # Gear ratio
+        motor_pos_current = joint_pos_current / self.gear_ratio
+        motor_vel_current = joint_vel_current / self.gear_ratio
+        motor_effort_current = joint_effort_current * self.gear_ratio
+        motor_pos_ref = joint_pos_ref / self.gear_ratio
+        motor_vel_ref = joint_vel_ref / self.gear_ratio
+        motor_effort_ref = joint_effort_ref * self.gear_ratio
 
         # Print rows (batch: all joints in ONE log block)
         header_str = (
             f"{'ID':>3}  {'NAME':<12}  "
-            f"{'POS':>15}  {'VEL':>15}  {'CMD':>15}  {'SAFE':>15}  {'REF':>15}"
+            f"{'POS':>15}  {'VEL':>12}  {'CMD_TAU':>12}  {'CMD_POS':>12}  {'CMD_VEL':>12}"
         )
         div_str = "-" * len(header_str)
-
         lines = [header_str, div_str]
 
         # Print log data
         for joint_index in range(self.num_joints):
             name = self.joint_names[joint_index] if joint_index < len(self.joint_names) else f"joint_{joint_index}"
 
-            raw_ref = float(ref_vector[joint_index])
-            raw_safe = float(safe_joint_targets[joint_index])
-            if joint_index in self.wheel_indices:
-                # Wheel: Reference is Speed
-                ref_val = float(np.rad2deg(raw_ref)) if self.print_degrees else raw_ref
-                ref_safe = float(np.rad2deg(raw_safe)) if self.print_degrees else raw_safe
-                ref_unit = velocity_unit
-            else:
-                # Joint: Reference is Position
-                ref_val = float(np.rad2deg(raw_ref)) if self.print_degrees else raw_ref
-                ref_safe = float(np.rad2deg(raw_safe)) if self.print_degrees else raw_safe
-                ref_unit = position_unit
-
-            # Integrate data
+            # Integrate data (all data is !!motor!! based)
             lines.append(
                 f"{joint_index:>3}  {name[:12]:<12}  "
-                f"{float(joint_position[joint_index]):>9.{self.precision}f} {position_unit:<5}  "
-                f"{float(joint_velocity[joint_index]):>9.{self.precision}f} {velocity_unit:<5}  "
-                f"{float(command_value[joint_index]):>9.{self.precision}f} {command_unit:<5}  "
-                f"{float(ref_safe):>9.{self.precision}f} {ref_unit:<5}  "
-                f"{float(ref_val):>9.{self.precision}f} {ref_unit:<5}"
+                f"{float(motor_pos_current[joint_index]):>9.{self.precision}f} {position_unit:<5}  "
+                f"{float(motor_vel_current[joint_index]):>9.{self.precision}f} {velocity_unit:<5}  "
+                f"{float(motor_effort_ref[joint_index]):>9.{self.precision}f} {self.command_unit:<5}  "
+                f"{float(motor_pos_ref[joint_index]):>9.{self.precision}f} {position_unit:<5}  "
+                f"{float(motor_vel_ref[joint_index]):>9.{self.precision}f} {velocity_unit:<5}"
             )
 
         # Leading newline: print one line below the logger prefix ([INFO] ...)
         self.get_logger().info("\n" + "\n".join(lines))
         self._print_count += 1
 
-
 def main(args=None):
     rclpy.init(args=args)
-    node = LogViewer()
+    node = LogViewerNode()
     try:
         rclpy.spin(node)
     except KeyboardInterrupt:
