+++ Realworld/src/goat_control/setup.py
@@ -39,23 +39,13 @@ setup(
     },
     entry_points={
         'console_scripts': [
-            # 'goat_control_node = goat_control.nodes.control_node:main',
-            # 'state_estimation_node = goat_control.nodes.state_estimation_node:main',
-            # 'motor_io_node = goat_control.nodes.motor_io_node:main',
-            # 'agent_node = goat_control.nodes.agent_node:main',
-            # 'log_viewer_node = goat_control.nodes.log_viewer_node:main',
-            # 'policy_keyboard_tester = goat_control.nodes.policy_keyboard_tester:main',
-            # 'plot_node = goat_control.nodes.plot_node:main',
-            # 'calibration_node = goat_control.nodes.calibration_node:main',
-            # 'nsc_tester_node = goat_control.nodes.nsc_tester_node:main',
-
-            'calibration_node = goat_control.nodes_develop.calibration_node:main',
-            'controller_node = goat_control.nodes_develop.controller_node:main',
-            'imu_io_node = goat_control.nodes_develop.imu_io_node:main',
-            'log_viewer_node = goat_control.nodes_develop.log_viewer_node:main',
-            'motor_io_node = goat_control.nodes_develop.motor_io_node:main',
-            'sim_controller_node = goat_control.nodes_develop.sim_controller_node:main',
-            'topic_converter_node = goat_control.nodes_develop.topic_converter_node:main',
+            'calibration_node = goat_control.nodes.calibration_node:main',
+            'controller_node = goat_control.nodes.controller_node:main',
+            'imu_io_node = goat_control.nodes.imu_io_node:main',
+            'log_viewer_node = goat_control.nodes.log_viewer_node:main',
+            'motor_io_node = goat_control.nodes.motor_io_node:main',
+            'sim_controller_node = goat_control.nodes.sim_controller_node:main',
+            'topic_converter_node = goat_control.nodes.topic_converter_node:main',
         ],
     },
 )
